change int to long in ARingZZp code.  Doesn't affect correctness.

+++ M2/Macaulay2/e/aring-zzp.cpp
@@ -82,7 +82,7 @@ namespace M2 {
                                bool p_plus,
                                bool p_parens) const
   {
-    int n = coerceToLongInteger(a);
+    long n = coerceToLongInteger(a);
     if (n < 0) 
       {
         o << '-';
@@ -95,7 +95,7 @@ namespace M2 {
 
   void ARingZZp::eval(const RingMap *map, const elem f, int first_var, ring_elem &result) const
   {
-    int a = coerceToLongInteger(f);
+    long a = coerceToLongInteger(f);
     result = map->get_ring()->from_long(a);
   }
